Allow users to set Logo URL and Data URI
This doesn't link up with the RootAction - as I couldn't figure out how
to load the data from the PageDecorator (I tried loading the descriptor,
but that didn't allow me to ge the data).

I would also  like to stop to using the NoOp theme to add a RootAction,
as this means all other themes need to inherit this functionality some
how. It would be better if I could add a RootAction which wasn't linked
to the theme - but it didn't seem to get called.